Introduce EmergencyResponderArchives, all tests pass.
Once an emergency is resolved, the responder is freed.
Calculations like full_response are no longer correct at this point.
We could move that onto the db, but the fact that into a boolean on the db,
but the fact that we are interested in things like full_response? implies that
at some point, we might want to know "well, how close was it?" or
"did we have firefighters sitting around listless, because the emergency required
medical attention?" which would show up as a not full response, but with enough
information to get an understanding of why (in this case: stop giving incentives
for graduates to go into firefighting, we have more than we are using, instead
move the incentives to medical responders, where we have a need).

This gets us really close to being able to rebuild the entire history
(the mutability of Responder fucks it up, as the value of on_duty is transient,
capacity could potentially change as they become better trained and gain experience).
If we can rebuild the entire history, though, then we can do things like
develop alternative algorithms and run them against the history of emergencies and
responders to see if they would have performed better or worse.

Other nice things fall out of this, as well: responders can now view their history,
we can see how well utilized responders are (eg is the responder with the most capacity
being used? or is that capacity usually surplus? Should we adjust the dispatch algorithm
to shoot for more balance? Or to select a dispatcher with a small surplus over 2 dispatchers
that match exactly, because having 2 responders available is more valuable than having
a few extra capacity. Questions like that can now be answered.